Fix the bug program flow away if bin is large
The bug is caused by wrong environment. When bin is large, in the function
board_init might call function which over than 4K before jump to SDRAM. So
the start should be careful to split bootloader to two step: bl0 just init
sdram, stack, nand and copy code to sdram, then the pc is set to SDRAM
address. So board_init is ran in SDRAM rather than stepping store of SRAM
which causes error.