Fix the bug program flow away in task context
The compiler would optimize the function call approach. The return
address of function would be like this:

1) if the called function do none function calls(it means the called
function would modify pc), the lr register is not pushed to stack;
2) if the called function do any function calls(it means the pc would
be modified), the lr register is pushed.

The context seems never would change the pc, while there is a swi ins.
This caused the lr flushed.

Using -O2 seems to solve this bug.